Refactor localized-data unit-test
